- Refactoring, order, and put some strings as a resource. - Added some tests for database performance, and for getting the elapsed time since a post was made.

+++ app/src/main/java/com/example/fernando/SimpleAPP/Post.java
@@ -3,14 +3,14 @@ package com.example.fernando.SimpleAPP;
 public class Post {
 
     private String ID;
-    private int Created_at;
+    private long Created_at;
     private String Author;
     private String Title;
     private String Url;
     private int Flag;
 
 
-    Post(int created_at, String author, String title, String url, String id, int flag) {
+    Post(long created_at, String author, String title, String url, String id, int flag) {
         this.Created_at = created_at;
         this.Author = author;
         this.Title = title;
@@ -19,7 +19,7 @@ public class Post {
         this.Flag = flag;
     }
 
-    public int getCreated_at() {
+    public long getCreated_at() {
         return Created_at;
     }
